requirements/requirements_armv7.txt
joblib==0.17.0                                                                                                                                                                                      
numpy==1.18.3                                                                                                                                                                                       
opencv-contrib-python==4.1.0.25                                                                                                                                                                     
pkg-resources==0.0.0                                                                                                                                                                                
PyYAML==5.3.1                                                                                                                                                                                       
scikit-learn==0.23.2                                                                                                                                                                                
scipy==1.5.1
ffmpeg-python
threadpoolctl==2.1.0